+++ SLIX/visualization.py
@@ -3,16 +3,20 @@ from matplotlib.colors import hsv_to_rgb
 from matplotlib import pyplot as plt
 from PIL import Image
 import copy
+import tqdm
 
-from SLIX._visualization import _downsample, _downsample_2d, _count_nonzero
+from SLIX._visualization import _downsample, _plot_axes_unit_vectors, \
+    _visualize_multiple_direction, \
+    _visualize_one_direction
 
-__all__ = ['visualize_parameter_map',
-           'visualize_unit_vectors',
-           'visualize_direction']
+__all__ = ['parameter_map',
+           'unit_vectors',
+           'unit_vector_distribution',
+           'direction']
 
 
-def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1,
-                            cmap='viridis', vmin=0, vmax=None, colorbar=True):
+def parameter_map(parameter_map, fig=None, ax=None, alpha=1,
+                  cmap='viridis', vmin=0, vmax=None, colorbar=True):
     """
     This method will create a Matplotlib plot based on imshow to display the
     given parameter map in different colors. The parameter map is plotted to
@@ -63,10 +67,10 @@ def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1,
     return fig, ax
 
 
-def visualize_unit_vectors(UnitX, UnitY, ax=None, thinout=20,
-                           scale=-1, vector_width=1,
-                           alpha=0.8, background_threshold=0.5,
-                           background_value=0):
+def unit_vectors(UnitX, UnitY, ax=None, thinout=20,
+                 scale=-1, vector_width=1,
+                 alpha=0.8, background_threshold=0.5,
+                 background_value=0):
     """
     This method will create a Matplotlib plot based on quiver to represent the
     given unit vectors as colored lines (vector map).
@@ -143,47 +147,132 @@ def visualize_unit_vectors(UnitX, UnitY, ax=None, thinout=20,
 
         for i in range(UnitX.shape[2]):
             UnitX[:, :, i] = numpy.array(
-                Image.fromarray(downscaled_unit_x[:, :, i]) \
-                .resize(UnitX.shape[:2][::-1], Image.NEAREST)
+                Image.fromarray(downscaled_unit_x[:, :, i])
+                     .resize(UnitX.shape[:2][::-1], Image.NEAREST)
             )
             UnitY[:, :, i] = numpy.array(
-                Image.fromarray(downscaled_unit_y[:, :, i]) \
-                .resize(UnitY.shape[:2][::-1], Image.NEAREST)
+                Image.fromarray(downscaled_unit_y[:, :, i])
+                     .resize(UnitY.shape[:2][::-1], Image.NEAREST)
             )
 
         del downscaled_unit_y
         del downscaled_unit_x
     for i in range(UnitX.shape[2]):
-        mesh_x, mesh_y = numpy.meshgrid(numpy.arange(0, UnitX.shape[1], thinout),
-                                        numpy.arange(0, UnitX.shape[0], thinout))
+        mesh_x, mesh_y = numpy.meshgrid(numpy.arange(0, UnitX.shape[1],
+                                                     thinout),
+                                        numpy.arange(0, UnitX.shape[0],
+                                                     thinout))
         mesh_u = UnitX[::thinout, ::thinout, i]
         mesh_v = UnitY[::thinout, ::thinout, i]
 
-        # Normalize the arrows:
-        mesh_u_normed = mesh_u / numpy.sqrt(numpy.maximum(1e-15,
-                                            mesh_u ** 2 + mesh_v ** 2))
-        mesh_v_normed = mesh_v / numpy.sqrt(numpy.maximum(1e-15,
-                                            mesh_u ** 2 + mesh_v ** 2))
-        mesh_u_normed[numpy.isclose(mesh_u, 0) &
-                      numpy.isclose(mesh_v, 0)] = numpy.nan
-        mesh_v_normed[numpy.isclose(mesh_u, 0) &
-                      numpy.isclose(mesh_v, 0)] = numpy.nan
-
-        normed_angle = numpy.arctan2(mesh_v_normed, -mesh_u_normed)
-        color_rgb = visualize_direction(normed_angle * 180.0 / numpy.pi)
-        color_rgb = numpy.clip(color_rgb.reshape(color_rgb.shape[0] *
-                                                 color_rgb.shape[1], 3), 0, 1)
-
-        # 1/scale to increase vector length for scale > 1
-        ax.quiver(mesh_x, mesh_y, mesh_u_normed, mesh_v_normed,
-                  color=color_rgb, angles='xy', scale_units='xy',
-                  scale=1.0/scale, headwidth=0, headlength=0, headaxislength=0,
-                  minlength=0, pivot='mid', alpha=alpha,
-                  width=vector_width, units='xy', edgecolors=color_rgb)
+        _plot_axes_unit_vectors(ax,
+                                mesh_x.flatten(),
+                                mesh_y.flatten(),
+                                mesh_u.flatten(),
+                                mesh_v.flatten(),
+                                scale, alpha, vector_width)
     return ax
 
 
-def visualize_direction(direction):
+def unit_vector_distribution(UnitX, UnitY, ax=None, thinout=20,
+                             scale=-1, vector_width=1,
+                             alpha=0.01):
+    """
+    This method will create a Matplotlib plot based on quiver to represent the
+    given unit vectors as colored lines (vector map).
+    Instead of showing a single vector like in unit_vector, here each vector
+    will be shown in the resulting image. The thinout parameter will determine
+    how many vectors will be overlapping. It is recommended to use a very small
+    alpha value to see which directions in the resulting plot are dominant.
+    Here, the vectors will only be plotted
+    to the current axis. To show the results, please use pyplot.show(). The
+    result might need some time to show depending on the input image size.
+
+    Args:
+
+        UnitX: Unit vector components along the x-axis (3D NumPy array).
+
+        UnitY: Unit vector components along the y-axis (3D NumPy array).
+
+        thinout: Downscaling parameter N (defines how many vectors N x N are
+        replaced by one vector).
+        Unit vectors will be thinned out using downscaling and thinning in
+        combination. This will increase the
+        vector size in the resulting image but will also reduce the information
+        density. Please use with caution.
+
+        scale: Increase the vector length by the given scale. Vectors will be
+               longer and might overlap if the scale is too high.
+
+        ax: Matplotlib axis. If None, the current context axis will be used.
+
+        vector_width: When choosing a high scale, the vectors might appear
+        quite thin which results in hard to read images. This option allows to
+        increase the vector thickness to improve visibility.
+
+        alpha: Apply alpha to Matplotlib plots to overlay them with some other
+        other image like the averaged transmitted light intensity.
+
+    Returns:
+
+        The current Matplotlib axis. The image can be shown with pyplot.show().
+
+    """
+    if ax is None:
+        ax = plt.gca()
+
+    while len(UnitX.shape) < 3:
+        UnitX = UnitX[..., numpy.newaxis]
+    while len(UnitY.shape) < 3:
+        UnitY = UnitY[..., numpy.newaxis]
+
+    # The default scale is below zero to allow the user to define his own scale
+    # A scale below zero isn't valid for visualization. If the user
+    # defines no scale, we suspect that the user wants an image
+    # where each vector has a scale of one. Therefore we set the scale to
+    # the same as our thinout when we draw the image.
+    if scale < 0:
+        scale = thinout
+
+    mesh_x = numpy.empty(UnitX.size)
+    mesh_y = numpy.empty(UnitX.size)
+    mesh_u = numpy.empty(UnitX.size)
+    mesh_v = numpy.empty(UnitX.size)
+    idx = 0
+
+    progress_bar = tqdm.tqdm(total=thinout*thinout,
+                             desc='Creating unit vectors.')
+    for offset_x in range(thinout):
+        for offset_y in range(thinout):
+            progress_bar.update(1)
+            for i in range(UnitX.shape[2]):
+                mesh_x_it, mesh_y_it = numpy.meshgrid(
+                    numpy.arange(0, UnitX.shape[1] - offset_x, thinout),
+                    numpy.arange(0, UnitX.shape[0] - offset_y, thinout)
+                )
+                mesh_x_it = mesh_x_it.flatten()
+                mesh_y_it = mesh_y_it.flatten()
+                mesh_u_it = UnitX[offset_y::thinout, offset_x::thinout, i]\
+                    .flatten()
+                mesh_v_it = UnitY[offset_y::thinout, offset_x::thinout, i]\
+                    .flatten()
+
+                mesh_x[idx:idx + len(mesh_x_it)] = mesh_x_it
+                mesh_y[idx:idx + len(mesh_y_it)] = mesh_y_it
+                mesh_u[idx:idx + len(mesh_u_it)] = mesh_u_it
+                mesh_v[idx:idx + len(mesh_v_it)] = mesh_v_it
+
+                idx = idx + len(mesh_x_it)
+
+    progress_bar.set_description('Finished. Plotting unit vectors.')
+    _plot_axes_unit_vectors(ax, mesh_x, mesh_y, mesh_u, mesh_v,
+                            scale, alpha, vector_width)
+    progress_bar.set_description('Done')
+    progress_bar.close()
+    return ax
+
+
+def direction(direction, saturation=None, value=None):
     """
     Generate a 2D colorized direction image in the HSV color space based on
     the original direction. Value and saturation of the color will always be
@@ -219,6 +308,14 @@ def visualize_direction(direction):
         direction: 2D or 3D Numpy array containing the direction of the image
                    stack
 
+        saturation: Weight image by using the saturation value. Use either a 2D image
+                    or a 3D image with the same shape as the direction. If no image
+                    is used, the saturation for all image pixels will be set to 1
+
+        value:  Weight image by using the value. Use either a 2D image
+                or a 3D image with the same shape as the direction. If no image
+                is used, the value for all image pixels will be set to 1
+
     Returns:
 
         numpy.ndarray: 2D image containing the resulting HSV orientation map
@@ -227,11 +324,30 @@ def visualize_direction(direction):
     direction = numpy.array(direction)
     direction_shape = direction.shape
 
-    h = direction
-    s = numpy.ones(direction.shape)
-    v = numpy.ones(direction.shape)
+    hue = direction
+    # If no saturation is given, create an "empty" saturation image that will be used
+    if saturation is None:
+        saturation = numpy.ones(direction.shape)
+    # Normalize saturation image
+    saturation = saturation / saturation.max()
+    # If we have a saturation image, check if the shape matches (3D) and correct accordingly
+    while len(saturation.shape) < len(direction.shape):
+        saturation = saturation[..., numpy.newaxis]
+    if not saturation.shape[-1] == direction_shape[-1]:
+        saturation = numpy.repeat(saturation, direction_shape[-1], axis=-1)
 
-    hsv_stack = numpy.stack((h / 180.0, s, v))
+    # If no value is given, create an "empty" value image that will be used
+    if value is None:
+        value = numpy.ones(direction.shape)
+    # Normalize value image
+    value = value / value.max()
+    # If we have a value image, check if the shape matches (3D) and correct accordingly
+    while len(value.shape) < len(direction.shape):
+        value = value[..., numpy.newaxis]
+    if not value.shape[-1] == direction_shape[-1]:
+        value = numpy.repeat(value, direction_shape[-1], axis=-1)
+
+    hsv_stack = numpy.stack((hue / 180.0, saturation, value))
     hsv_stack = numpy.moveaxis(hsv_stack, 0, -1)
     rgb_stack = hsv_to_rgb(hsv_stack)
 
@@ -239,64 +355,3 @@ def visualize_direction(direction):
         return _visualize_multiple_direction(direction, rgb_stack)
     else:
         return _visualize_one_direction(direction, rgb_stack)
-
-
-def _visualize_one_direction(direction, rgb_stack):
-    output_image = rgb_stack
-    output_image[direction == -1] = 0
-
-    return output_image.astype('float32')
-
-
-def _visualize_multiple_direction(direction, rgb_stack):
-    output_image = numpy.zeros((direction.shape[0] * 2,
-                                direction.shape[1] * 2,
-                                3))
-    # count valid directions
-    valid_directions = numpy.count_nonzero(direction > -1, axis=-1)
-
-    r = rgb_stack[..., 0]
-    g = rgb_stack[..., 1]
-    b = rgb_stack[..., 2]
-
-    # Now we need to place them in the right pixel on our output image
-    for x in range(direction.shape[0]):
-        for y in range(direction.shape[1]):
-            if valid_directions[x, y] == 0:
-                output_image[x * 2:x * 2 + 2, y * 2:y * 2 + 2] = 0
-            elif valid_directions[x, y] == 1:
-                output_image[x * 2:x * 2 + 2, y * 2:y * 2 + 2, 0] = r[x, y, 0]
-                output_image[x * 2:x * 2 + 2, y * 2:y * 2 + 2, 1] = g[x, y, 0]
-                output_image[x * 2:x * 2 + 2, y * 2:y * 2 + 2, 2] = b[x, y, 0]
-            else:
-                output_image[x * 2, y * 2, 0] = r[x, y, 0]
-                output_image[x * 2, y * 2, 1] = g[x, y, 0]
-                output_image[x * 2, y * 2, 2] = b[x, y, 0]
-
-                output_image[x * 2 + 1, y * 2, 0] = r[x, y, 1]
-                output_image[x * 2 + 1, y * 2, 1] = g[x, y, 1]
-                output_image[x * 2 + 1, y * 2, 2] = b[x, y, 1]
-
-                if valid_directions[x, y] == 2:
-                    output_image[x * 2, y * 2 + 1, 0] = r[x, y, 1]
-                    output_image[x * 2, y * 2 + 1, 1] = g[x, y, 1]
-                    output_image[x * 2, y * 2 + 1, 2] = b[x, y, 1]
-
-                    output_image[x * 2 + 1, y * 2 + 1, 0] = r[x, y, 0]
-                    output_image[x * 2 + 1, y * 2 + 1, 1] = g[x, y, 0]
-                    output_image[x * 2 + 1, y * 2 + 1, 2] = b[x, y, 0]
-                else:
-                    output_image[x * 2, y * 2 + 1, 0] = r[x, y, 2]
-                    output_image[x * 2, y * 2 + 1, 1] = g[x, y, 2]
-                    output_image[x * 2, y * 2 + 1, 2] = b[x, y, 2]
-
-                    if valid_directions[x, y] == 3:
-                        output_image[x * 2 + 1, y * 2 + 1, 0] = 0
-                        output_image[x * 2 + 1, y * 2 + 1, 1] = 0
-                        output_image[x * 2 + 1, y * 2 + 1, 2] = 0
-                    if valid_directions[x, y] == 4:
-                        output_image[x * 2 + 1, y * 2 + 1, 0] = r[x, y, 3]
-                        output_image[x * 2 + 1, y * 2 + 1, 1] = g[x, y, 3]
-                        output_image[x * 2 + 1, y * 2 + 1, 2] = b[x, y, 3]
-
-    return output_image.astype('float32')