+++ bin/SLIXParameterGenerator.py
@@ -4,7 +4,7 @@ import sys
 sys.path.append(os.getcwd())
 sys.path.append("..")
 
-from SLIX._cmd import main_full_image
+from SLIX._cmd.ParameterGenerator import main
 if __name__ == '__main__':
     sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
-    sys.exit(main_full_image())
+    sys.exit(main())