+++ bin/SLIXVisualizeParameter.py
@@ -4,7 +4,7 @@ import sys
 sys.path.append(os.getcwd())
 sys.path.append("..")
 
-from SLIX._cmd import main_visualize
+from SLIX._cmd.VisualizeParameter import main
 if __name__ == '__main__':
     sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
-    sys.exit(main_visualize())
+    sys.exit(main())